+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Fix issue where webhook data passed to `LocalPaymentReversed` webhook class is missing
+* Add `grantedPaymentInstrumentUpdate` to `WebhookNotification`
 
 ## 3.5.0
 * Add `paymentReaderCardDetails` parameter to `Transaction.sale`